2/1/2020 Updates
Added the Flash Blossom Reindeer cards

Fixed some older cards with outdated texts

Fixed some of the old code for Hazmat Animal (least for Gazelle and the Field's old codes)

Please note that the game won't render out the prompts in CDB; likely due to the 9 digit thing. You may wish to have it open on the side or take notes. They should generally be ordered in the same places they appear on text.