ED0Pro updating
As YGOPro will no longer be updated due to ED0Pro, I am shifting gears.

The pics and arch files should be the same, though code will be different in some places.

Please let me know if any of the code is busted in some places so I can fix the problem.

(Read the Update log)